camera: bsa: Fix releasing of threads
Threads in buffer source adapter were not being released correctly,
which lead to them being orphaned. Previous logic relied on the
Thread destructors getting called to wake up the threads for them
to exit, but the destructors were not getting called since it still
had child threads.

This patch implements the virtual requestExit function for each
Thread class. Each implementation wakes its respective thread if
it is waiting, so the thread can exit.

Change-Id: I95c42a05ce76d6241b073c5927ecb81365562066
Signed-off-by: Tyler Luu <tluu@ti.com>
Signed-off-by: Vladimir Petrov <vppetrov@mm-sol.com>